Remove refetching of rates
This patch removes the manual refetching of rates via the runtime
messaging system. I chose to remove it as it made the flow of data very
hard to follow (as evident by the fact that the current implementation
contained multiple bugs). Now, we just aim to fail gracefully instead of
halfhartedly attempting to update the rates upon page load.

Additionally, this path also updates the background script to use the
new promise API for every storage access + adds some comments.