Don't clear rates on failure
This patch removes some of the behavior added by fb94d03. Deleting the
old rates upon failure was a bit _too_ aggressive. The rationale for
this patch is as follows:

Image that the API is down for a couple of hours. Since we also removed
refetching of rates, that would leave a user with no rates at all for 24
hours. With this new patch they will just be running the old rates until
the next scheduled update.